% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_survival_aux.R
\name{alphaRevSurv}
\alias{alphaRevSurv}
\title{Estimate alpha parameter for mortality}
\usage{
alphaRevSurv(lx_std, qx, type)
}
\arguments{
\item{lx_std}{standard values for mortality estimation}

\item{qx}{3 element vector of child mortality (q0_5) or women adult mortality (q15_45) data}

\item{type}{'child' if estimating values for children or 'women' for adult women}
}
\value{
alpha, vector of three coefficient elements for 0-4, 5-9 and 10-14 years prior reference date
}
\description{
Estimate alpha parameter for mortality of past years basedon provided qx and
lx_standard parameters for women and children
}
\keyword{internal}
