% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceb.R
\name{ceb_eval}
\alias{ceb_eval}
\title{Cohort analysis of mean children ever born (CEB)}
\usage{
ceb_eval(date_svy, ages_w, pop_w, ceb, mac = 28, plot_ceb = TRUE)
}
\arguments{
\item{date_svy}{survey or census date in the format Y-m-d or string in the formats
Y, Y-m, Y-m-d}

\item{ages_w}{women starting age of 5 year age group}

\item{pop_w}{total women matching ages in ages_w}

\item{ceb}{total children ever born by women age group}

\item{mac}{mean age at childbearing matching ages in age_w or unique value for all ages,
default = 28}

\item{plot_ceb}{dummy variable to return the plot for mean CEB results}
}
\value{
data.frame with 2 vectors: reference date of mean children ever born (decimal)
and mean children ever born
}
\description{
Construct estimates of mean CEB in the reference period of childbearing
(mean age of childbearing)
}
\examples{

## Kenya 1989 census
date_svy <- "1989-08-25"
ages_w <- seq( 40, 75, 5 )
pop_w <- c( 350140, 280920, 230080, 173260, 158140, 111360, 82080, 54220 )
ceb <- c( 2532140, 2151920, 1736540, 1314140, 1143740, 816820, 560520, 371060 )

ceb_eval( date_svy, ages_w, pop_w, ceb, mac = 28, plot_ceb = TRUE )
}
