% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parity_assess.R
\name{coh_parassess}
\alias{coh_parassess}
\title{Cohort Parity Assessment function}
\usage{
coh_parassess(avg_par1, avg_par2, time_span = 10)
}
\arguments{
\item{avg_par1}{data.frame with at least two columns $ages (vector of ages or starting ages of age group intervals) and $P
(vector of average parities by women age group) for census/survey year 1}

\item{avg_par2}{data.frame with at least two columns $ages (vector of ages or starting ages of age group intervals) and $P
(vector of average parities by women age group) for census/survey year 2}

\item{time_span}{census/survey time span (default = FALSE)}
}
\value{

}
\description{
Cohort Parity Assessment function
}
\examples{

}
