% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.prty_KHM}
\alias{data.prty_KHM}
\title{Parity data from Cambodia 2008 Census}
\format{
A data frame with 3 variables:
  parity (from 0 to N and including NA response codes 98, 99 or NA for example),
  ages (starting age group interval)
  women (women counts of respective parity from respective age group)
}
\source{
Moultrie TA, RE Dorrington, AG Hill, K Hill, IM Timæus and B Zaba (eds). 2013.
  Tools for Demographic Estimation. Paris: International Union for the Scientific Study of
  Population. demographicestimation.iussp.org
}
\usage{
data.prty_KHM
}
\description{
Parity data from Cambodia 2008 Census
}
\keyword{datasets}
