% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_survival_aux.R
\name{decimal_anydate}
\alias{decimal_anydate}
\title{Converts a date of multiple formats to decimal}
\usage{
decimal_anydate(date)
}
\arguments{
\item{date}{date in the format Y-m-d or string in the formats Y, Y-m, Y-m-d}
}
\value{
decimal of date's year
}
\description{
Converts a date of multiple formats (year, year-month, year-month-day) to decimal
}
\examples{
date1 <- '2011'
# converts date1 for decimal value of '2011-07-31'
decimal_anydate(date1)

date2 <- '2009-05'
# converts date2 for decimal value of '2009-05-15'
decimal_anydate(date2)

date3 <- '2013-10-31'
# converts date3 for decimal value of '2013-10-31'
decimal_anydate(date3)
}
