% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brass_pf.R
\name{fertBrassPF.cltrss}
\alias{fertBrassPF.cltrss}
\title{Brass PF Fertility Estimation using Coale-Trussel coefficients}
\source{
United Nations. 1983. Manual X: Indirect techniques for demographic estimation
(United Nations publication, Sales No. E.83.XIII.2).
}
\usage{
fertBrassPF.cltrss(ages = seq(15, 45, 5), P, asfr,
  adjust_group = c(20), age_shift = TRUE)
}
\arguments{
\item{ages}{A vector of starting ages of five-year age groups ranging from 15 to 45 (default = c(15,20,25,30,35,40,45))}

\item{P}{A vector of mean parities by five-year age group - same groups as 'ages'}

\item{asfr}{A vector of age-specific fertility rates by five-year age group - same groups as 'ages'}

\item{adjust_group}{A vector of age-groups from 'ages' to be used for selection of PF ratios to adjust asfr data
(default set to 20 (20-24 five-year age group))}

\item{age_shift}{TRUE (default) if fertility rates are calculated from births in a 12-month period
by age of mothed at the end of period and FALSE if fertility rates are calculated from births by
age of mother at the end of the period}
}
\value{
A list with 3 elements:
pf_data data frame with columns ages, P for mean parities, asfr, Fi for cumulate fertility estimated from Brass coefficients,PF for ratios P/F and adj_asfr for adjusted asfr;
tfr_unadj for unadjusted total fertility rate estimate;
and tfr_adj for adjusted total fertility rate estimate by applying the selected age-group PF ratio
}
\description{
Brass PF Fertility Estimation using Coale-Trussel coefficients
}
\examples{
## Bangladesh 1974 survey data:
ages_bd = c(15, 20, 25, 30, 35, 40, 45)
asfr_bd = c(0.1063, 0.2296, 0.2154, 0.1825, 0.1339, 0.0644, 0.0336)
P_bd    = c(0.385, 1.847, 3.485, 4.917, 5.861, 6.194, 6.084)
fertBrassPF.cltrss(P = P_bd, asfr = asfr_bd, adjust_group = c(20,25,30))


}
