% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brass_pf.R
\name{fertBrassPF}
\alias{fertBrassPF}
\title{Brass PF Fertility Estimation}
\usage{
fertBrassPF(ages = seq(15, 45, 5), P, asfr, adjust_group = c(20))
}
\arguments{
\item{ages}{A vector of starting ages of five-year age groups ranging from 15 to 45 (default = c(15,20,25,30,35,40,45))}

\item{P}{A vector of mean parities by five-year age group - same groups as 'ages'}

\item{asfr}{A vector of age-specific fertility rates by five-year age group - same groups as 'ages'}

\item{adjust_group}{A vector of age-groups from 'ages' to be used for selection of PF ratios to adjust asfr data
(default set to 20 (20-24 five-year age group))}
}
\value{
A list with 3 elements:
pf_data data frame with columns ages, P for mean parities, asfr, Fi for cumulate fertility estimated from Brass coefficients,PF for ratios P/F and adj_asfr for adjusted asfr;
tfr_unadj for unadjusted total fertility rate estimate;
and tfr_adj for adjusted total fertility rate estimate by applying the selected age-group PF ratio
}
\description{
Brass PF Fertility Estimation
}
\examples{
## Malawi 2008 Census data:
ages_ma = c(15, 20, 25, 30, 35, 40, 45)
asfr_ma = c(0.111, 0.245, 0.230, 0.195, 0.147, 0.072, 0.032)
P_ma    = c(0.283, 1.532, 2.849, 4.185, 5.214, 6.034, 6.453)
fertBrassPF(P = P_ma, asfr = asfr_ma)


}
