% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_survival_clean.R
\name{FertRevSurvWPP}
\alias{FertRevSurvWPP}
\title{Reverse Survival Fertility Estimation Function for WPP2019 data}
\source{
Moultrie TA, RE Dorrington, AG Hill, K Hill, IM Timæus and B Zaba (eds). 2013.
  Tools for Demographic Estimation. Paris: International Union for the Scientific Study of
  Population. demographicestimation.iussp.org
}
\usage{
FertRevSurvWPP(locations, date_ref, lt_family = "General", logquad = FALSE)
}
\arguments{
\item{locations}{list of locations by name or code according to WPP 2019 location list (check fertestr::locs_avail() for list of available locations)}

\item{date_ref}{reference date of inquiry given in the following formats:
Y-m-d (4 digit year - 2 digit month - 2 digit day), Y-m (4 digit year - 2 digit month),
Y (4 digit year)}

\item{lt_family}{model life table family to be used to retrieve single age survival information}

\item{logquad}{if TRUE estimates lx functions from logquad models instead of using lt_family (default FALSE)}
}
\value{
data.frame with 5 elements:
`location_code`: WPP 2019 location code
`location_name`: WPP 2019 location name
`year` - reference period of fertility estimation in decimal format;
`TFR`  - estimated total fertility rate;
`births` - estimated total number of births
}
\description{
Apply reverse survival fertility estimation to WPP2019 country data
}
\examples{

countries <- c( 32, 76, 380, 508, 752 ) # Argentina, Brazil, Italy, Mozambique, Sweden
FertRevSurvWPP( locations = countries, date_ref = 2010, lt_family = 'General' )


}
