% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_survival_aux.R
\name{fetch_FertPattern_Wpp2019}
\alias{fetch_FertPattern_Wpp2019}
\title{Function to get fertility pattern from Wpp 2019 data}
\usage{
fetch_FertPattern_Wpp2019(country_code = NULL, year)
}
\arguments{
\item{country_code}{list of country codes to retrieve fertility pattern data from}

\item{year}{period of reference to retrieve fertility pattern data}
}
\value{
data.frame with three columns \item{age}: women ages, \item{asfr_std_ref}:
age-specific fertility rates for the time-period which contains the reference year and
\item{asfr_std_15prior}: age-specific fertility rates for the time-period
15 years prior to the reference period
}
\description{
Function to get fertility pattern from Wpp 2019 data
}
\keyword{internal}
