% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceb.R
\name{fetch_mac_Wpp2019}
\alias{fetch_mac_Wpp2019}
\title{Function to retrieve mean age of childbearing from Wpp 2019 data}
\usage{
fetch_mac_Wpp2019(country_code = NULL, year)
}
\arguments{
\item{country_code}{list of country codes to retrieve fertility pattern data from}

\item{year}{period of reference to retrieve fertility pattern data}
}
\value{
mean age at childbearing for country and year selected
}
\description{
Function to retrieve mean age of childbearing from Wpp 2019 data
}
\keyword{internal}
