% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_survival_aux.R
\name{fetch_MortProb_Wpp2019}
\alias{fetch_MortProb_Wpp2019}
\title{Function to get child and adult-women mortality from Wpp 2019 data}
\usage{
fetch_MortProb_Wpp2019(country_code = NULL, year)
}
\arguments{
\item{country_code}{list of country codes to retrieve mortality pattern data from}

\item{year}{period of reference to retrieve mortality}
}
\description{
Function to get child and adult-women mortality from Wpp 2019 data
}
\keyword{internal}
