% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_survival_aux.R
\name{FetchLifeTableWpp2019}
\alias{FetchLifeTableWpp2019}
\title{Retrieve life table functions from WPP 2019 data}
\usage{
FetchLifeTableWpp2019(locations = NULL, year, sex = "both")
}
\arguments{
\item{locations}{list of location codes or names to retrieve LT information}

\item{year}{period of reference to generate estimates}

\item{sex}{sex to retrieve mortality (male, female or both)}
}
\description{
Retrieve life tables for selected period and locations from WPP 2019 data
}
