% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_survival_aux.R
\name{FetchPopWpp2019}
\alias{FetchPopWpp2019}
\title{Fetch population data from WPP2019}
\usage{
FetchPopWpp2019(locations = NULL, year, ages, age_interval = 1, sex = "total")
}
\arguments{
\item{locations}{location id from WPP 2019 data}

\item{year}{period of reference of population data}

\item{ages}{selected ages to retrieve pop data}

\item{age_interval}{how to display ages in result - single ages (1 - default)
or 5-year age group (5)}

\item{sex}{sex to retrieve information form (default - total)}
}
\description{
Fetch population data from WPP2019
}
\keyword{internal}
