% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpp_fetch.R
\name{get_location_code}
\alias{get_location_code}
\title{Get WPP 2019 country codes from country names}
\usage{
get_location_code(location_names = NULL)
}
\arguments{
\item{location_names}{location name or vector of location names}
}
\value{
WPP 2019 `location_code` for given location names
}
\description{
Provides the list of countries and respective codes available in WPP 2019 or
fetch the country code for given country name
}
\examples{
# provides all location codes
get_location_code()

# provides location codes for a given list of countries
names <- c('Brazil','Argentina','Uruguay','Paraguay')
get_location_code( names )

}
