% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_survival_aux.R
\name{get_mlt}
\alias{get_mlt}
\title{Get lx data from model life tables}
\usage{
get_mlt(lt_family, e0, ages, sex)
}
\arguments{
\item{lt_family}{Model Life Table family name (Chilean, Far_East_Asian, Latin, General, South_Asian, North, South, East, West)}

\item{e0}{Life expectancy level for life table (lower bound = 20)}

\item{ages}{age selection of data (single age-interval from 0 to 130)}

\item{sex}{`Female` or `Male`}
}
\value{
data.frame with selected ages `$age` and survival functions `$lx_std`
}
\description{
Retrieve survival data for provided model life tables
}
\keyword{internal}
