% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_survival_aux.R
\name{interpolate}
\alias{interpolate}
\title{2 point interpolation}
\usage{
interpolate(y1, y2, x1, x2, x)
}
\arguments{
\item{y1, y2, x1, x2}{Reference known points (x1,y1), (x2,y2)}

\item{x}{Point for which we want to estimate y based on slope given by points 1 and 2}
}
\value{
y = (x - x1)/(x2-x1) * (y2-y1) + y1
}
\description{
2 point interpolation
}
\keyword{internal}
