% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpp_fetch.R
\name{is_LocID}
\alias{is_LocID}
\title{Check if given location code corresponds to a location ID}
\usage{
is_LocID(location)
}
\arguments{
\item{location}{}
}
\value{
TRUE if location corresponds to a location ID
}
\description{
Check if given location code corresponds to a location ID
}
\keyword{internal}
