% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ken_par_1989}
\alias{ken_par_1989}
\title{Parity data from Kenya 1989 Census}
\format{A data frame with 8 variables:
  parity (from 0 to 30 and including NA responses),
  PX (number of women reffered to specific parity of age group X, X from 1:15-19 to 7:45-49)}
\source{
Moultrie TA, RE Dorrington, AG Hill, K Hill, IM Timæus and B Zaba (eds). 2013.
  Tools for Demographic Estimation. Paris: International Union for the Scientific Study of
  Population. demographicestimation.iussp.org
}
\usage{
ken_par_1989
}
\description{
Parity data from Kenya 1989 Census
}
\keyword{datasets}
