% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_named_data}
\alias{load_named_data}
\title{Loads dataset from package}
\usage{
load_named_data(name_data, pkg)
}
\arguments{
\item{name_data}{name of the data as a string}

\item{pkg}{name of the package where the data is as a string}
}
\description{
Loads dataset from package
}
