% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_survival_aux.R
\name{locs_avail}
\alias{locs_avail}
\title{Available locations from WPP 2019 data}
\usage{
locs_avail()
}
\value{
data.frame with two columns name and location_code
}
\description{
Provides the list of available locations from WPP 2019 data
}
\examples{
# available countries from wpp2019
locs_avail()

}
