% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_survival_aux.R
\name{logit}
\alias{logit}
\title{Logit lx}
\usage{
logit(lx = NULL, qx = NULL)
}
\arguments{
\item{lx}{survival function}

\item{qx}{probability of deaths}
}
\value{
Yx, survival function logit value
}
\description{
Cumpute logit of survival function from lx or qx
}
\keyword{internal}
