% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{modelLTx1}
\alias{modelLTx1}
\title{Model Life Tables United Nations}
\format{
A data frame with 15 columns:
    source (model life table source: 'CD' for Coale-Demeny or 'UN' for Uniter Nations)
    family (model life table family: 'Chilean', 'Far_East_Asian', 'General, 'Latin', 'South_Asian',
    'East', 'West', 'North', 'South')
    type_mlt (source + family name)
    sex ('Female' or 'Male')
    e0 (life expectancy level)
    and other life table functions for one year age interval
    age, mx1, qx1, lx1, dx1, Lx1, Tx1, sx1, ex1, ax1
}
\source{
Extended Model Life Tables - United Nations 2011
  https://www.un.org/en/development/desa/population/publications/mortality/model-life-tables.asp
}
\usage{
modelLTx1
}
\description{
Model Life Tables United Nations for one year age interval
}
\keyword{datasets}
