% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mult.age_shift}
\alias{mult.age_shift}
\title{Coefficients to estimate age-specific fertility rates for conventional age groups}
\format{
A data frame with age coefficients x, y and z
}
\source{
United Nations. 1983. Manual X: Indirect techniques for demographic estimation
(United Nations publication, Sales No. E.83.XIII.2).
}
\usage{
mult.age_shift
}
\description{
Coefficients for calculation of weighting factors to estimate age-specific fertility rates for
conventional age groups shifted by six months
}
\keyword{datasets}
