% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parity_assess.R
\name{parity_assessment}
\alias{parity_assessment}
\title{Parity Assessment function}
\usage{
parity_assessment(avg_par, par_assess_graph = FALSE)
}
\arguments{
\item{avg_par}{data.frame with at least two columns $ages (vector of ages or starting ages of age group intervals) and $P
(vector of average parities by women age group)}

\item{par_assess_graph}{TRUE of FALSE for parity assessment diagnose plot output (default = FALSE)}
}
\value{
A warning message if the average parities are not monotonically increasing by age group and the
original avg_par data.frame with an additional column $diff_P with the inter-cohort average parity differences
}
\description{
Parity Assessment function
}
\examples{
## Malawi 2008 average parity:
parity_assessment(avg_par = mwi_fert_2008, par_assess_graph = T)
}
