% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parity_assess.R
\name{prtyAssess.cohort}
\alias{prtyAssess.cohort}
\title{Parity Assessment function between two surveys}
\usage{
prtyAssess.cohort(ages.t1, ages.t2, P.t1, P.t2, time.span = 10)
}
\arguments{
\item{ages.t1}{vector of ages or starting ages of age group intervals for census/survey year 1}

\item{ages.t2}{vector of ages or starting ages of age group intervals for census/survey year 2}

\item{P.t1}{vector of average parities by women age group for census/survey year 1}

\item{P.t2}{vector of average parities by women age group for census/survey year 2}

\item{time.span}{census/survey time span (default = 10)}
}
\value{

}
\description{
Verify if the average number of children ever born of women birth cohorts increase between two surveys/censuses
}
\examples{

}
