% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parity_assess.R
\name{prtyAssess.plot}
\alias{prtyAssess.plot}
\title{Parity Assessment Plot}
\source{
Moultrie TA, RE Dorrington, AG Hill, K Hill, IM Timæus and B Zaba (eds). 2013.
  Tools for Demographic Estimation. Paris: International Union for the Scientific Study of
  Population. demographicestimation.iussp.org
}
\usage{
prtyAssess.plot(ages, P)
}
\arguments{
\item{ages}{A vector of ages or starting ages of age group intervals}

\item{P}{A vector of average parities by women age group}
}
\value{
Average parities vs ages plot
}
\description{
Parity Assessment Plot
}
\examples{

library(DemoToolsData)

### Malawi 2008 data:
prtyAssess.plot( ages = data.pf_MWI$ages, P = data.pf_MWI$P )

}
