% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parity_assess.R
\name{prtyAssess}
\alias{prtyAssess}
\title{Parity Assessment  Function}
\usage{
prtyAssess(ages, P, prtyAssess.graph = FALSE)
}
\arguments{
\item{ages}{A vector of ages or starting ages of age group intervals}

\item{P}{A vector of average parities by women age group}

\item{prtyAssess.graph}{TRUE of FALSE for El-Badry diagnose plot output (default = FALSE)}
}
\value{
A warning message if the average parities are not monotonically increasing by age group and the
original avg_par data.frame with an additional column $diff_P with the inter-cohort average parity differences
}
\description{
Parity Assessment  Function
}
\examples{
### Malawi 2008 data:
prtyAssess( ages = data.pf_MWI$ages, P = data.pf_MWI$P )
}
