% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parity_assess.R
\name{prtyAverage}
\alias{prtyAverage}
\title{Average Parity Computation Function}
\usage{
prtyAverage(ages, parity, women, na_code = NA, eb_cor = FALSE,
  impl_par = FALSE, eb_graph = FALSE, par_assess = FALSE,
  par_assess_graph = FALSE, age_group = "q")
}
\arguments{
\item{ages}{A vector of ages or starting ages of age group intervals}

\item{parity}{A vector of parities for each age group}

\item{women}{A vector of women counts by parity and age group}

\item{na_code}{A numeric value representing the label for missing parities (default = NA)}

\item{eb_cor}{TRUE or FALSE for correction of zeros and missing values by El-Badry method (default = FALSE)}

\item{impl_par}{TRUE or FALSE for correction of implausible parities (default = FALSE)}

\item{eb_graph}{TRUE of FALSE for El-Badry diagnose plot output (default = FALSE)}

\item{par_assess}{TRUE or FALSE for parity assessment of average parity results (default = FALSE)}

\item{par_assess_graph}{TRUE or FALSE for parity assessment plot output (default = FALSE)}

\item{age_group}{Character assuming values 'q' (default) for quinquennial age group or 's' for single age group input}
}
\value{
A data.frame with 2 variables: ages and P for average parities by age group
}
\description{
Average Parity Computation Function
}
\examples{
## Kenya 1989 data:
avg.parity(data_par = ken_par_1989)
# With El-Badry:
avg.parity(data_par = ken_par_1989, eb_cor = TRUE, eb_graph = TRUE)
# Correction for implausible parities
prtyAverage(data_par = ken_par_1989, impl_par = TRUE)
}
