% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parity_assess.R
\name{prtyElBadry}
\alias{prtyElBadry}
\title{El-Badry correction Function}
\usage{
prtyElBadry(ages, parity, women, prtyElBadry.graph = FALSE,
  na_code = NA)
}
\arguments{
\item{ages}{A vector of ages or starting ages of age group intervals}

\item{parity}{A vector of parities for each age group}

\item{women}{A vector of women counts by parity and age group}

\item{prtyElBadry.graph}{TRUE of FALSE for El-Badry diagnose plot output (default = FALSE)}

\item{na_code}{A numeric value representing the code for missing parities (default = NA)}
}
\value{
The original data_par data.frame with corrected numbers of zero and missing parities
estimated by the El-Badry function
}
\description{
El-Badry correction Function
}
\examples{
### Kenya 1989 data:
prtyElBadry( ages = data.prty_KEN$ages, parity = data.prty_KEN$parity, women = data.prty_KEN$women, prtyElBadry.graph = T )
###
### Cambodia 2008 data:
prtyElBadry( ages = data.prty_KHM$ages, parity = data.prty_KHM$parity, women = data.prty_KHM$women )
###
}
