% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parity_assess.R
\name{prtyImplaus}
\alias{prtyImplaus}
\title{Implausible Parities correction function}
\usage{
prtyImplaus(ages, parity, women, age_group = "q")
}
\arguments{
\item{ages}{A vector of ages or starting ages of age group intervals}

\item{parity}{A vector of parities for each age group}

\item{women}{A vector of women counts by parity and age group}

\item{age_group}{Character assuming values 'q' (default) for quinquennial
age group or 's' for single age group input}
}
\value{
The original data_par data.frame with implausible parities set to 0
}
\description{
Implausible Parities correction function
}
\examples{
### Kenya 1989 data:
prtyImplaus( ages = data.prty_KEN$ages, parity = data.prty_KEN$parity, women = data.prty_KEN$women )
}
