% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_survival_aux.R
\name{q_calcWpp2019}
\alias{q_calcWpp2019}
\title{Compute qx probability of death}
\usage{
q_calcWpp2019(location_code, years, sex, age_inf, age_sup)
}
\arguments{
\item{location_code}{location code from WPP 2019 data}

\item{years}{years to retrieve estimation of qx}

\item{sex}{female, male or both}

\item{age_inf}{lower bound age of estimation}

\item{age_sup}{upper_bound age of estimation}
}
\value{
data.frame with 3 columns, years of estimation, age_interval, qx
}
\description{
Calculate qx probability of death between age interval age_inf, age_sup for WPP 2019 locations
}
\keyword{internal}
