% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_survival_aux.R
\name{revSurvMainx5}
\alias{revSurvMainx5}
\title{Reverse Survival Estimation for 5-year age groups of children}
\usage{
revSurvMainx5(year, datWomen, lxWomen_std, q15_45f, fertPattern, datChildren)
}
\arguments{
\item{year}{reference date of estimation in decimal format}

\item{lxWomen_std}{female survival functions data.frame for reverse survival of females}

\item{q15_45f}{female adult mortality probability 3 element vector or single value}

\item{fertPattern}{female fertility pattern (age-specific standardized rates)}

\item{datChildren}{children population data.frame}

\item{datWoman}{women population data.frame}
}
\value{
estimates of TFR by year prior to reference date
}
\description{
Estimate TFR levels from processed information of women, children and date for five year periods before inquiry
}
\keyword{internal}
