% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_survival_aux.R
\name{revSurvMainx5_Beta}
\alias{revSurvMainx5_Beta}
\title{Reverse Survival Estimation for 5-year age group of children (BETA)}
\usage{
revSurvMainx5_Beta(
  year,
  datWomen,
  lxWomen_std,
  q15_45f,
  fertPattern,
  datChildren,
  lxChildren_std,
  q0_5,
  a0_5,
  a5_5,
  a10_5
)
}
\arguments{
\item{year}{reference date of estimation in decimal format}

\item{lxWomen_std}{female survival functions data.frame for reverse survival of females}

\item{q15_45f}{female adult mortality probability 3 element vector or single value}

\item{fertPattern}{female fertility pattern (age-specific standardized rates)}

\item{datChildren}{children population data.frame}

\item{lxChildren_std}{children survival functions data.frame for reverse survival of
children}

\item{q0_5}{children mortality probability 3 element vector or single value}

\item{a0_5}{a0_5}

\item{a5_5}{a5_5}

\item{a10_5}{a10_5}

\item{datWoman}{women population data.frame}
}
\value{
estimates of TFR by year prior to reference date
}
\description{
Estimate TFR levels from processed information of women, children and date
}
\keyword{internal}
