% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{std.zaba}
\alias{std.zaba}
\title{Zaba standard table for Gompertz model fertility estimation}
\format{}
\source{
Moultrie TA, RE Dorrington, AG Hill, K Hill, IM Timæus and B Zaba (eds). 2013.
  Tools for Demographic Estimation. Paris: International Union for the Scientific Study of
  Population. demographicestimation.iussp.org
}
\usage{
std.zaba
}
\description{
Zaba standard table for Gompertz model fertility estimation
}
\keyword{datasets}
