% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_survival_aux.R
\name{womenRevSurv}
\alias{womenRevSurv}
\title{Estimate women survival probabilities}
\usage{
womenRevSurv(age, lx_std, women, alphaWomen, year, std_asfr)
}
\arguments{
\item{age}{age vector related to standard survival values and women population}

\item{lx_std}{standard survival function for selected women}

\item{women}{women population for selected ages}

\item{alphaWomen}{alpha parameters estimated from alphaRevSurv function}

\item{year}{reference period of estimation}

\item{std_asfr}{fertility pattern - standardized age-fertility rates}
}
\value{
data.frame with parameters:
year: year of reverse survived population
AgesWomen: women reproductive period ages
popWomen: women population
asfr_std: standardized age-specific fetrtility rates for estimation period
}
\description{
Estimate women survival probabilities for females aged 15-49 the reference date
}
\keyword{internal}
