% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gomp_pf.R
\name{fertGompPF}
\alias{fertGompPF}
\title{Gompertz PF Fertility Estimation}
\source{
Brass W, AJ. 1968. Coale Methods  of  analysis  and  estimation.  In:  BRASS,  W.  et  al.  (Ed.).  The demography of tropical Africa. 1. ed. New Jersey: Princeton University Press, p. 88-139.
Brass W. 1975. Methods for Estimating Fertility and Mortality from Limited and Defected Data. North Carolina: Carolina Population Center.
}
\usage{
fertGompPF(ages, asfr, P, age_shift = TRUE, graph_check = F,
  rmse_check = T)
}
\arguments{
\item{ages}{A vector of starting ages of five-year age groups ranging from 15 to 45 (default = c(15,20,25,30,35,40,45))}

\item{asfr}{A vector of age-specific fertility rates by five-year age group - same groups as 'ages'}

\item{P}{A vector of mean parities by five-year age group - same groups as 'ages'}

\item{adjust_group}{A vector of age-groups from 'ages' to be used for selection of PF ratios to adjust asfr data
(default set to 20 (20-24 five-year age group))}
}
\value{
A list with 3 elements:
pf_data data frame with columns ages, P for mean parities, asfr, Fi for cumulate fertility estimated from Brass coefficients,PF for ratios P/F and adj_asfr for adjusted asfr;
tfr_unadj for unadjusted total fertility rate estimate;
and tfr_adj for adjusted total fertility rate estimate by applying the selected age-group PF ratio
}
\description{
Gompertz PF Fertility Estimation
}
\examples{
## Malawi 2008 Census data:
ages_ma = c(15, 20, 25, 30, 35, 40, 45)
asfr_ma = c(0.111, 0.245, 0.230, 0.195, 0.147, 0.072, 0.032)
P_ma    = c(0.283, 1.532, 2.849, 4.185, 5.214, 6.034, 6.453)
fertBrassPF(P = P_ma, asfr = asfr_ma)


}
