% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_survival_clean.R
\name{FertRevSurvx5c}
\alias{FertRevSurvx5c}
\title{Reverse Survival Fertility Estimation Function (x5)}
\source{
Moultrie TA, RE Dorrington, AG Hill, K Hill, IM Timæus and B Zaba (eds). 2013.
  Tools for Demographic Estimation. Paris: International Union for the Scientific Study of
  Population. demographicestimation.iussp.org
}
\usage{
FertRevSurvx5c(
  ages5_c = c(0, 5, 10),
  popx5_c,
  ages5_w = seq(10, 65, 5),
  popx5_w,
  lx1_c,
  lx5_w,
  asfr = c(0, 0.017, 0.055, 0.057, 0.041, 0.022, 0.007, 0.002),
  asfr_15prior = NULL,
  q0_5 = NULL,
  q15_45f = NULL,
  date_ref
)
}
\arguments{
\item{ages5_c}{children five-year age group ages vector (default 0, 5, 10)}

\item{popx5_c}{children population in five-year age group (x5) matching ages5_c vector}

\item{ages5_w}{women five-year age group ages vector ( default seq( 10, 65, 5 ) )}

\item{popx5_w}{women population in five-year age group format (x5) matching ages_w vector}

\item{lx1_c}{children survival function vector in single age groups from 0 to 15}

\item{lx5_w}{women survival function vector in five-year age groups matching ages_w vector}

\item{asfr}{age specific fertility rates for five-year age groups from 10-45 for current period
of estimation}

\item{asfr_15prior}{standardized age specific fertility rates for five-year age groups from 10-45 for the
period of 15 years before the current inquiry period}

\item{q0_5}{3 element vector for mortality probability between ages 0-4 for the period of estimation,
period 5 years prior to estimation period, and period 10 years prior to estimation period}

\item{q15_45f}{female adult mortality probability for the period of estimation,
period 5 years prior to estimation period, and period 10 years prior to estimation period}

\item{date_ref}{reference date of inquiry given in the following formats:
Y-m-d (4 digit year - 2 digit month - 2 digit day), Y-m (4 digit year - 2 digit month),
Y (4 digit year)}
}
\value{
data.frame with 3 elements:
`year` - reference period of fertility estimation in decimal format;
`TFR`  - estimated total fertility rate;
`births` - estimated total number of births
}
\description{
Reverse Survival Fertility Estimation for five-year age groups of children
}
\examples{

# 1 - User input data
popx5_c <-  c( 1376790, 1472470, 1673110 )

popx5_w <- c(  815930, 780320, 697160, 626430, 361650, 435880,
               393760, 352520, 294280, 230200, 160590, NA )

lx1_c <- c( 1.0000, 0.9320, 0.9275, 0.9228, 0.9165, 0.9125, 0.9110,
            0.9094, 0.9079, 0.9063, 0.9048, 0.9032, 0.9017, 0.9001,
            0.8986, 0.8970 )

lx5_w <- c( 0.91381, 0.90989, 0.90492, 0.89798, 0.88893, 0.87596,
            0.86029, 0.84188, 0.81791, 0.78472, 0.73735, 0.67316 )

q0_5 <-  c( 0.0683, 0.1008, 0.1189)

q15_45f <- c( 0.1946, 0.2290, 0.2674 )

asfr <- c( 0.0000, 0.0418,0.1535, 0.1482, 0.1118, 0.0708,
           0.0301, 0.0032 )

asfr_15prior <- c( 0.0000, 0.0533, 0.1974, 0.2144, 0.1836, 0.1332,
                   0.0676, 0.0134 )

FertRevSurvx5c( ages5_c = c( 0, 5, 10 ), popx5_c = popx5_c,
                ages5_w = seq( 10, 65, 5 ), popx5_w = popx5_w,
                lx1_c = lx1_c, lx5_w = lx5_w,
                asfr = asfr,
                asfr_15prior = asfr_15prior,
                q0_5 = q0_5, q15_45f = q15_45f,
                date_ref = '2008-03-03' )


}
