% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_survival_aux.R
\name{FetchFertilityWpp2019}
\alias{FetchFertilityWpp2019}
\title{Retrieve Fertility Pattern from WPP 2019}
\usage{
FetchFertilityWpp2019(locations = NULL, year)
}
\arguments{
\item{locations}{}

\item{year}{}
}
\value{
data.frame with three columns age: women ages, asfr_std_ref:
age-specific fertility rates for the time-period which contains the reference year and
asfr_std_15prior: age-specific fertility rates for the time-period
15 years prior to the reference period
}
\description{
Retrieve age-specific fertility rates for available WPP 2019 locations
}
