% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_survival_aux.R
\name{get_location_name}
\alias{get_location_name}
\title{Get WPP 2019 location name from codes}
\usage{
get_location_name(country_code)
}
\arguments{
\item{country_code}{country code}
}
\value{
country_name
}
\description{
Get WPP 2019 location name from codes
}
\keyword{internal}
