% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpp_fetch.R
\name{get_location_name}
\alias{get_location_name}
\title{Get WPP 2019 location name from codes}
\usage{
get_location_name(location_code)
}
\arguments{
\item{location_code}{WPP 2019 location code}
}
\value{
`location_name` of given `location_code`
}
\description{
Get WPP 2019 location name from codes
}
\examples{
# Get name for Argentina (code 32)
get_location_name( 32 )

# Get name for list of countries
get_location_name( c( 76, 32, 600, 604, 152 ))

}
