% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parity_assess.R
\name{implausib_parity}
\alias{implausib_parity}
\title{Implausible Parities correction function}
\usage{
implausib_parity(data_par, age_group)
}
\arguments{
\item{data_par}{data.frame with columns $ages (vector of ages or starting ages of age group intervals),
$parity (vector of parities for each age group) and $women (vector of women counts by parity and age group)}

\item{age_group}{Character assuming values 'q' (default) for quinquennial
age group or 's' for single age group input}

\item{women}{A vector of women counts by parity and age group}
}
\value{
The original data_par data.frame with implausible parities set to 0
}
\description{
Implausible Parities correction function
}
\examples{
## Kenya 1989 data:
implausib_parity(data_par = ken_par_1989)
}
