% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{khm_par_2008}
\alias{khm_par_2008}
\title{Parity data from Cambodia 2008 Census}
\format{A data frame with 9 variables:
  parity (from 0 to 30 and including NA responses),
  PX (number of women reffered to specific parity of age group X, X from 0:10-14 to 7:45-49)}
\source{
Moultrie TA, RE Dorrington, AG Hill, K Hill, IM Timæus and B Zaba (eds). 2013.
  Tools for Demographic Estimation. Paris: International Union for the Scientific Study of
  Population. demographicestimation.iussp.org
}
\usage{
khm_par_2008
}
\description{
Parity data from Cambodia 2008 Census
}
\keyword{datasets}
