% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parity_assess.R
\name{parity_assessment}
\alias{parity_assessment}
\title{El-Badry correction Function}
\usage{
parity_assessment(avg_par, par_assess_graph = FALSE)
}
\arguments{
\item{avg_par}{data.frame with two columns $ages (vector of ages or starting ages of age group intervals) and $P
(vector of average parities by women age group)}

\item{eb_graph}{TRUE of FALSE for El-Badry diagnose plot output (default = FALSE)}
}
\value{
The original data_par data.frame with corrected numbers of zero and missing parities
estimated by the El-Badry function
}
\description{
El-Badry correction Function
}
\examples{
## Kenya 1989 data:
el_badry(data_par = ken_par_1989, eb_graph = T)
}
