% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parity_assess.R
\name{prtyElBadry}
\alias{prtyElBadry}
\title{El-Badry correction Function}
\usage{
prtyElBadry(ages, parity, women, eb_graph = FALSE, na_code = NA)
}
\arguments{
\item{ages}{A vector of ages or starting ages of age group intervals}

\item{parity}{A vector of parities for each age group}

\item{women}{A vector of women counts by parity and age group}

\item{eb_graph}{TRUE of FALSE for El-Badry diagnose plot output (default = FALSE)}

\item{na_code}{A numeric value representing the code for missing parities (default = NA)}
}
\value{
The original data_par data.frame with corrected numbers of zero and missing parities
estimated by the El-Badry function
}
\description{
El-Badry correction Function
}
\examples{
## Kenya 1989 data:
prtyElBadry(data_par = ken_par_1989, eb_graph = T)
}
