% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parity_assess.R
\name{prtyElBadry}
\alias{prtyElBadry}
\title{El-Badry correction Function}
\source{
Moultrie TA, RE Dorrington, AG Hill, K Hill, IM Timæus and B Zaba (eds). 2013.
Tools for Demographic Estimation. Paris: International Union for the Scientific Study of
Population. demographicestimation.iussp.org

el-Badry MA. 1961.
"Failure of enumerators to make entries of zero: errors in recording childless cases in population censuses",
 Journal of the American Statistical Association 56(296):909–924.
 doi: http://dx.doi.org/10.1080/01621459.1961.10482134

 UN Population Division. 1983. Manual X: Indirect Techniques for Demographic Estimation.
 New York: United Nations, Department of Economic and Social Affairs, ST/ESA/SER.A/81.
 http://www.un.org/esa/population/techcoop/DemEst/manual10/manual10.html
}
\usage{
prtyElBadry(ages, parity, women, prtyElBadry.graph = FALSE, na_code = NA)
}
\arguments{
\item{ages}{A vector of ages or starting ages of age group intervals}

\item{parity}{A vector of parities for each age group}

\item{women}{A vector of women counts by parity and age group}

\item{prtyElBadry.graph}{TRUE of FALSE for El-Badry diagnose plot output (default = FALSE)}

\item{na_code}{A numeric value representing the code for missing parities (default = NA)}
}
\value{
A list with 3 elements: $data_par data.frame with corrected numbers of zero and missing parities
estimated by the El-Badry function, $prty.Zi data.frame with proportion of zero parities and $prty.Ui with
proportion of unkown parities
}
\description{
El-Badry correction Function
}
\examples{
### Kenya 1989 data:
prtyElBadry( ages = data.prty_KEN$ages, parity = data.prty_KEN$parity, women = data.prty_KEN$women, prtyElBadry.graph = T )
###
### Cambodia 2008 data:
prtyElBadry( ages = data.prty_KHM$ages, parity = data.prty_KHM$parity, women = data.prty_KHM$women )
###
}
