% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parity_assess.R
\name{prtyImplaus}
\alias{prtyImplaus}
\title{Implausible Parities correction Function}
\usage{
prtyImplaus(ages, parity, women, age_group = "q")
}
\arguments{
\item{ages}{A vector of ages or starting ages of age group intervals}

\item{parity}{A vector of parities for each age group}

\item{women}{A vector of women counts by parity and age group}

\item{age_group}{Character assuming values 'q' (default) for quinquennial
age group or 's' for single age group input}
}
\value{
The original data_par data.frame with implausible parities set to 0
}
\description{
Implausible Parities correction Function
}
\examples{
## Kenya 1989 data:
prtyImplaus(data_par = ken_par_1989)
}
